+++ src/LaravelDomainsFacade.php
@@ -5,15 +5,12 @@ namespace Pushberryfinn\LaravelDomains;
 use Illuminate\Support\Facades\Facade;
 
 /**
- * @see \Pushberryfinn\LaravelDomains\Skeleton\SkeletonClass
+ * @see \Pushberryfinn\LaravelDomains\LaravelDomains
+ *
+ * @method static array domains()
  */
 class LaravelDomainsFacade extends Facade
 {
-    /**
-     * Get the registered name of the component.
-     *
-     * @return string
-     */
     protected static function getFacadeAccessor()
     {
         return 'laravel-domains';