+++ src/LaravelDomainsServiceProvider.php
@@ -2,19 +2,28 @@
 
 namespace Pushberryfinn\LaravelDomains;
 
-use Illuminate\Support\ServiceProvider;
 use Illuminate\Support\Facades\File;
+use Illuminate\Support\ServiceProvider;
 use Illuminate\Support\Str;
 
 class LaravelDomainsServiceProvider extends ServiceProvider
 {
+    public function register()
+    {
+        $this->mergeConfigFrom(__DIR__ . '/../config/config.php', 'laravel-domains');
+
+        $this->app->singleton('laravel-domains', fn () => new LaravelDomains());
+    }
 
     public function boot()
     {
         if ($this->app->runningInConsole()) {
+            $this->publishes([
+                __DIR__ . '/../config/config.php' => config_path('laravel-domains.php'),
+            ], 'laravel-domains-config');
+
             $this->registerConsoleCommands();
         }
-
     }
 
     protected function registerConsoleCommands()